@@ -65,17 +65,17 @@
         "@mui/icons-material": "^5.4.4",
         "@mui/material": "^5.4.3",
         "@testing-library/jest-dom": "^5.16.2",
-        "@testing-library/react": "^12.1.3",
+        "@testing-library/react": "^13.0.0",
         "@testing-library/react-hooks": "^7.0.2",
-        "@testing-library/user-event": "^13.5.0",
+        "@testing-library/user-event": "^14.0.4",
         "axios": "^0.26.0",
         "buffer": "^6.0.3",
         "date-fns": "^2.28.0",
         "prettier": "^2.5.1",
         "prop-types": "^15.8.1",
-        "react": "^17.0.2",
+        "react": "^18.0.0",
         "react-cookie": "^4.1.1",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.0.0",
         "react-icons": "^4.3.1",
         "react-portal": "^4.2.2",
         "react-responsive-modal": "^6.1.0",
@@ -83,7 +83,7 @@
         "react-router-dom": "^6.2.1",
         "react-scripts": "5.0.0",
         "react-spinners": "^0.11.0",
-        "react-test-renderer": "^17.0.2",
+        "react-test-renderer": "^18.0.0",
         "recoil": "^0.7.0",
         "styled-components": "^5.3.1",
         "uuid": "^8.3.2",
